Rename
renaming file name based on functionality
